Improvements to the limelight targeting code.
Compute the drive speeds for the three axes before commanding the drive.
In the previous implementation the drive was getting written to hardware
multiple times which could cause some drive issues.

There were also some potential bugs that could cause distance correction
to not be performed if the crosshair was aligned correctly in X.

Fixed some warnings in the robot class.